solution/api/pyproject.toml
[project]
name = "solution"
version = "0.1.0"
description = "PROD Advertising Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "better-profanity>=0.7.0",
    "django>=5.1.6",
    "django-clickhouse-backend>=1.3.2",
    "django-minio-storage>=0.5.7",
    "djangorestframework>=3.15.2",
    "faker>=36.1.1",
    "gunicorn>=23.0.0",
    "locust>=2.32.10",
    "numpy>=2.2.3",
    "pathlib>=1.0.1",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
    "yandex-cloud-ml-sdk>=0.3.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]
